+++ src/main/java/corecord/dev/common/config/SecurityConfig.java
@@ -4,8 +4,8 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import corecord.dev.common.response.ApiResponse;
 import corecord.dev.common.status.ErrorStatus;
 import corecord.dev.common.util.*;
-import corecord.dev.domain.auth.application.OAuthLoginFailureHandler;
-import corecord.dev.domain.auth.application.OAuthLoginSuccessHandler;
+import corecord.dev.domain.auth.handler.OAuthLoginFailureHandler;
+import corecord.dev.domain.auth.handler.OAuthLoginSuccessHandler;
 import lombok.RequiredArgsConstructor;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
