+++ src/main/java/corecord/dev/domain/auth/repository/TmpTokenRepository.java
@@ -1,6 +1,6 @@
-package corecord.dev.domain.token.repository;
+package corecord.dev.domain.auth.repository;
 
-import corecord.dev.domain.token.entity.TmpToken;
+import corecord.dev.domain.auth.entity.TmpToken;
 import org.springframework.data.repository.CrudRepository;
 
 import java.util.Optional;
