+++ src/main/java/corecord/dev/domain/user/service/UserService.java
@@ -5,10 +5,10 @@ import corecord.dev.common.status.ErrorStatus;
 import corecord.dev.common.util.CookieUtil;
 import corecord.dev.common.util.JwtUtil;
 import corecord.dev.domain.record.repository.RecordRepository;
-import corecord.dev.domain.token.entity.RefreshToken;
-import corecord.dev.domain.token.exception.enums.TokenErrorStatus;
-import corecord.dev.domain.token.exception.model.TokenException;
-import corecord.dev.domain.token.repository.RefreshTokenRepository;
+import corecord.dev.domain.auth.entity.RefreshToken;
+import corecord.dev.domain.auth.exception.enums.TokenErrorStatus;
+import corecord.dev.domain.auth.exception.model.TokenException;
+import corecord.dev.domain.auth.repository.RefreshTokenRepository;
 import corecord.dev.domain.user.converter.UserConverter;
 import corecord.dev.domain.user.dto.request.UserRequest;
 import corecord.dev.domain.user.dto.response.UserResponse;
